warnings.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/branches/ruby_1_9_3@34718 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ ext/date/date_core.c
@@ -2381,7 +2381,7 @@ offset_to_sec(VALUE vof, int *rof)
 	    n = FIX2LONG(vs);
 	    if (n < -DAY_IN_SECONDS || n > DAY_IN_SECONDS)
 		return 0;
-	    *rof = n;
+	    *rof = (int)n;
 	    return 1;
 	}
     }
